Engine: merged sprite.cpp contents into draw.cpp
