+++ Solutions/Engine.App/Engine.App.vcxproj.filters
@@ -354,9 +354,6 @@
     <ClCompile Include="..\..\Engine\ac\speech.cpp">
       <Filter>Source Files\ac</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Engine\ac\sprite.cpp">
-      <Filter>Source Files\ac</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\Engine\ac\string.cpp">
       <Filter>Source Files\ac</Filter>
     </ClCompile>
@@ -981,9 +978,6 @@
       <Filter>Header Files\ac</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Engine\ac\speech.h">
-      <Filter>Header Files\ac</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\Engine\ac\sprite.h">
       <Filter>Header Files\ac</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Engine\ac\string.h">
